LPS-180956 Extract COREntryQualifiersScreenNavigationCategory methods to a COREntryQualifiersScreenNavigationEntry component and extend COREntryQualifiersScreenNavigationCategory in it
