+++ modules/apps/commerce/commerce-order-rule-web/src/main/java/com/liferay/commerce/order/rule/web/internal/frontend/taglib/servlet/taglib/COREntryQualifiersScreenNavigationCategory.java
@@ -14,35 +14,13 @@
 
 package com.liferay.commerce.order.rule.web.internal.frontend.taglib.servlet.taglib;
 
-import com.liferay.commerce.currency.service.CommerceCurrencyService;
-import com.liferay.commerce.order.rule.entry.type.COREntryTypeJSPContributorRegistry;
-import com.liferay.commerce.order.rule.entry.type.COREntryTypeRegistry;
-import com.liferay.commerce.order.rule.model.COREntry;
-import com.liferay.commerce.order.rule.service.COREntryRelService;
-import com.liferay.commerce.order.rule.service.COREntryService;
-import com.liferay.commerce.order.rule.web.internal.display.context.COREntryQualifiersDisplayContext;
 import com.liferay.commerce.order.rule.web.internal.entry.constants.COREntryScreenNavigationEntryConstants;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -51,22 +29,14 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=20",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=20",
+	service = ScreenNavigationCategory.class
 )
 public class COREntryQualifiersScreenNavigationCategory
-	implements ScreenNavigationCategory, ScreenNavigationEntry<COREntry> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
-		return COREntryScreenNavigationEntryConstants.CATEGORY_KEY_QUALIFIERS;
-	}
-
-	@Override
-	public String getEntryKey() {
 		return COREntryScreenNavigationEntryConstants.CATEGORY_KEY_QUALIFIERS;
 	}
 
@@ -75,7 +45,7 @@ public class COREntryQualifiersScreenNavigationCategory
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(resourceBundle, "eligibility");
+		return language.get(resourceBundle, "eligibility");
 	}
 
 	@Override
@@ -84,80 +54,7 @@ public class COREntryQualifiersScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_COR_ENTRY_GENERAL;
 	}
 
-	@Override
-	public boolean isVisible(User user, COREntry corEntry) {
-		if (corEntry == null) {
-			return false;
-		}
-
-		boolean hasPermission = false;
-
-		try {
-			hasPermission = _corEntryModelResourcePermission.contains(
-				PermissionThreadLocal.getPermissionChecker(),
-				corEntry.getCOREntryId(), ActionKeys.UPDATE);
-		}
-		catch (Exception exception) {
-			if (_log.isDebugEnabled()) {
-				_log.debug(exception);
-			}
-		}
-
-		return hasPermission;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		COREntryQualifiersDisplayContext corEntryQualifiersDisplayContext =
-			new COREntryQualifiersDisplayContext(
-				_commerceCurrencyService, _corEntryModelResourcePermission,
-				_corEntryRelService, _corEntryService,
-				_corEntryTypeJSPContributorRegistry, _corEntryTypeRegistry,
-				httpServletRequest, _portal);
-
-		httpServletRequest.setAttribute(
-			WebKeys.PORTLET_DISPLAY_CONTEXT, corEntryQualifiersDisplayContext);
-
-		_jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse,
-			"/cor_entry/qualifiers.jsp");
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		COREntryQualifiersScreenNavigationCategory.class);
-
 	@Reference
-	private CommerceCurrencyService _commerceCurrencyService;
-
-	@Reference(
-		target = "(model.class.name=com.liferay.commerce.order.rule.model.COREntry)"
-	)
-	private ModelResourcePermission<COREntry> _corEntryModelResourcePermission;
-
-	@Reference
-	private COREntryRelService _corEntryRelService;
-
-	@Reference
-	private COREntryService _corEntryService;
-
-	@Reference
-	private COREntryTypeJSPContributorRegistry
-		_corEntryTypeJSPContributorRegistry;
-
-	@Reference
-	private COREntryTypeRegistry _corEntryTypeRegistry;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference
-	private Portal _portal;
+	protected Language language;
 
 }
